LF-11633: Segmentation fault on voice_ui_app after update libvoiceseeker to v0.6.11
When the libvoiceseeker is update to v0.6.11 the voice_ui_app stops
working with a Segmentation fault on rdspVoiceSpot_CreateControl ():

Program received signal SIGSEGV, Segmentation fault.
0x0000aaaaaaaad0b0 in rdspVoiceSpot_CreateControl ()

This error is caused by a change on the RdspDeviceConfig.h file.
This file is allocated in:
voiceseeker/platforms/iMXxxxxxxxxxxx/rdsp_utilities_public/include/
and
voicespot/platforms/iMXxxxxxxxxxxx/rdsp_utilities_public/include/

The make file for the voicespot must point to the path of the voicespot,
but is pointing to the voiceseeker path, so, is taking a change that is
not contemplated for voicespot.

Conclusion:
To avoid this Segmentation fault fix the path on the Makefile of
voicespot

Signed-off-by: Emmanuel Sanchez <emmanuel.sanchez@nxp.com>